+++ R/dialog.R
@@ -67,12 +67,7 @@ bc_dialog <- function(id,
       data_size = "icon-sm",
       `aria-label` = "Close dialog",
       onclick = paste0("this.closest('dialog').close()"),
-      HTML(
-        '<svg class="lucide lucide-x" xmlns="http://www.w3.org/2000/svg" width="24" ',
-        'height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" ',
-        'stroke-width="2" stroke-linecap="round" stroke-linejoin="round">',
-        '<path d="M18 6 6 18"/><path d="m6 6 12 12"/></svg>'
-      )
+      bc_icon("x")
     )
   )